+++ 1.21/Common/src/main/java/fuzs/diagonalblocks/DiagonalBlocks.java
@@ -7,6 +7,7 @@ import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.api.core.v1.ModLoaderEnvironment;
 import fuzs.puzzleslib.api.core.v1.utility.ResourceLocationHelper;
 import fuzs.puzzleslib.api.event.v1.core.EventPhase;
+import fuzs.puzzleslib.api.event.v1.entity.player.PlayerInteractEvents;
 import fuzs.puzzleslib.api.event.v1.server.TagsUpdatedCallback;
 import net.minecraft.resources.ResourceLocation;
 import org.apache.logging.log4j.LogManager;
@@ -29,6 +30,7 @@ public class DiagonalBlocks implements ModConstructor {
 
     private static void registerEventHandlers() {
         TagsUpdatedCallback.EVENT.register(EventPhase.FIRST, DiagonalBlockHandler::onTagsUpdated);
+        PlayerInteractEvents.USE_BLOCK.register(DiagonalBlockHandler::onUseBlock);
     }
 
     public static ResourceLocation id(String path) {