+++ 1.21/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.1-1.21] - 2024-08-12
+### Changed
+- Sneak + right-clicking with an empty hand now allows for toggling between diagonal and non-diagonal blocks
+### Fixed
+- Fix blocks not breaking faster with the correct tool
+
 ## [v21.0.0-1.21] - 2024-07-07
 - Port to Minecraft 1.21
 - Forge is no longer support in favor of NeoForge